# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Yingyan Platform API
  version: 0.1.0
  description: |
    映研用户端、管理端与 Worker 共用的 HTTP 契约。
    所有 JSON 成功响应均为 `{ code: 0, data }`，失败响应均为
    `{ code, message, details? }`。素材地址均为短期签名地址。
servers:
  - url: http://127.0.0.1:8080
    description: Local API
...
paths:
...
  /api/manage/redemption-batches:
    get:
      tags: [Manage Redemption]
      operationId: listRedemptionBatches
      x-permission: platform:manage
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: productCode
          in: query
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/ManageRedemptionBatchPage"
    post:
      tags: [Manage Redemption]
      operationId: createRedemptionBatch
      x-permission: platform:manage
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/RedemptionBatchCreate"
      responses:
        "201":
          $ref: "#/components/responses/ManageRedemptionBatchCreated"
        "409":
          $ref: "#/components/responses/Error"
  /api/manage/redemption-batches/{batchId}:
    get:
      tags: [Manage Redemption]
      operationId: getRedemptionBatch
      x-permission: platform:manage
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/BatchId"
      responses:
        "200":
          $ref: "#/components/responses/ManageRedemptionBatch"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Manage Redemption]
      operationId: updateRedemptionBatch
      description: 只修改批次展示名称，不改变批次内兑换码、次数、商品标识或有效期。
      x-permission: platform:manage
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/BatchId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/RedemptionBatchUpdate"
      responses:
        "200":
          $ref: "#/components/responses/ManageRedemptionBatch"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/manage/redemption-batches/{batchId}/reveal:
    post:
      tags: [Manage Redemption]
      operationId: revealRedemptionBatch
      x-permission: platform:manage
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/BatchId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/ManageGeneratedCodeList"
  /api/manage/redemption-batches/{batchId}/export:
    post:
      tags: [Manage Redemption]
      operationId: exportRedemptionBatch
      x-permission: platform:manage
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/BatchId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: format
          in: query
          description: csv 为运营表格，xianyu 为每码一行的闲鱼自动发货库存。
          schema:
            type: string
            enum: [csv, xianyu]
            default: csv
      responses:
        "200":
          $ref: "#/components/responses/ManageRedemptionBatchExport"
  /api/manage/redemption-codes:
    get:
      tags: [Manage Redemption]
      operationId: listRedemptionCodes
      x-permission: platform:manage
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RedemptionCodeStatus"
        - name: batchId
          in: query
          schema: { type: string, format: uuid }
        - name: productCode
          in: query
          schema: { type: string }
        - name: redeemedBy
          in: query
          schema: { type: string }
        - name: expiringSoon
          in: query
          schema: { type: boolean }
      responses:
        "200":
          $ref: "#/components/responses/ManageRedemptionCodePage"
  /api/manage/redemption-codes/{codeId}:
    get:
      tags: [Manage Redemption]
      operationId: getRedemptionCode
      x-permission: platform:manage
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CodeId"
      responses:
        "200":
          $ref: "#/components/responses/ManageRedemptionCodeDetail"
        "404":
          $ref: "#/components/responses/Error"
  /api/manage/redemption-codes/{codeId}/reveal:
    post:
      tags: [Manage Redemption]
      operationId: revealRedemptionCode
      x-permission: platform:manage
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/CodeId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/ManageRedemptionCodeReveal"
  /api/manage/redemption-codes/disable:
    post:
      tags: [Manage Redemption]
      operationId: disableRedemptionCodes
      x-permission: platform:manage
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/RedemptionCodesDisable"
      responses:
        "200":
          $ref: "#/components/responses/ManageBulkMutation"
  /api/manage/redemption-codes/extend:
    post:
      tags: [Manage Redemption]
      operationId: extendRedemptionCodes
      x-permission: platform:manage
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/RedemptionCodesExtend"
      responses:
        "200":
          $ref: "#/components/responses/ManageBulkMutation"
...
components:
  securitySchemes:
    userSession:
      type: apiKey
      in: cookie
      name: yy_user_session
    adminSession:
      type: apiKey
      in: cookie
      name: yy_manage_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      name: keyword
      in: query
      schema: { type: string, maxLength: 200 }
...
    BatchId:
      name: batchId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CodeId:
      name: codeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
...
    RedemptionBatchCreate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RedemptionBatchCreateInput"
    RedemptionBatchUpdate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RedemptionBatchUpdateInput"
    RedemptionCodesDisable:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [codeIds, reason]
            properties:
              codeIds:
                type: array
                minItems: 1
                maxItems: 500
                uniqueItems: true
                items: { type: string, format: uuid }
              reason: { type: string, minLength: 1, maxLength: 500 }
    RedemptionCodesExtend:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [codeIds, expiresAt, reason]
            properties:
              codeIds:
                type: array
                minItems: 1
                maxItems: 500
                uniqueItems: true
                items: { type: string, format: uuid }
              expiresAt: { type: string, format: date-time }
              reason: { type: string, minLength: 1, maxLength: 500 }
...
  responses:
...
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    ManageRedemptionBatchPage:
      description: 兑换码批次分页结果
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageRedemptionBatchPageSuccess"
    ManageRedemptionBatch:
      description: 兑换码批次
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageRedemptionBatchSuccess"
    ManageRedemptionBatchCreated:
      description: 新建兑换码批次及完整兑换码
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageRedemptionBatchCreatedSuccess"
    ManageGeneratedCodeList:
      description: 批次内仍可使用的完整兑换码
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageGeneratedCodeListSuccess"
    ManageRedemptionBatchExport:
      description: 兑换码批次 CSV 或闲鱼库存内容
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageRedemptionBatchExportSuccess"
    ManageRedemptionCodePage:
      description: 兑换码分页结果
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageRedemptionCodePageSuccess"
    ManageRedemptionCodeDetail:
      description: 兑换码及操作历史
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageRedemptionCodeDetailSuccess"
    ManageRedemptionCodeReveal:
      description: 单个完整兑换码
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageRedemptionCodeRevealSuccess"
    ManageBulkMutation:
      description: 兑换码批量操作结果
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManageBulkMutationSuccess"
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: integer, minimum: 1 }
        message: { type: string }
        details: true
...
    ApiSuccessEnvelope:
      type: object
      required: [code]
      properties:
        code: { type: integer, const: 0 }
    ManagePageMetadata:
      type: object
      required: [page, pageSize, total, hasMore]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        hasMore: { type: boolean }
...
    ManageRedemptionCounts:
      type: object
      additionalProperties: false
      required: [unused, redeemed, expired, disabled]
      properties:
        unused: { type: integer, minimum: 0 }
        redeemed: { type: integer, minimum: 0 }
        expired: { type: integer, minimum: 0 }
        disabled: { type: integer, minimum: 0 }
    ManageRedemptionBatch:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - productCode
        - quantity
        - creditsPerCode
        - expiresAt
        - neverExpires
        - createdBy
        - createdAt
        - counts
        - usageRate
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 60 }
        productCode: { type: string }
        quantity: { type: integer, minimum: 1, maximum: 500 }
        creditsPerCode: { type: integer, minimum: 1 }
        expiresAt: { type: [string, "null"], format: date-time }
        neverExpires: { type: boolean }
        note: { type: string }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        counts:
          $ref: "#/components/schemas/ManageRedemptionCounts"
        usageRate: { type: number, minimum: 0, maximum: 1 }
    ManageRedemptionCode:
      type: object
      additionalProperties: false
      required:
        - id
        - maskedCode
        - batchId
        - batchName
        - productCode
        - credits
        - status
        - expiresAt
        - createdAt
        - expiringSoon
      properties:
        id: { type: string, format: uuid }
        maskedCode: { type: string }
        batchId: { type: string, format: uuid }
        batchName: { type: string }
        productCode: { type: string }
        credits: { type: integer, minimum: 1 }
        status:
          $ref: "#/components/schemas/RedemptionCodeStatus"
        expiresAt: { type: [string, "null"], format: date-time }
        redeemedBy: { type: string, format: uuid }
        redeemedByEmail: { type: string, format: email }
        redeemedAt: { type: string, format: date-time }
        disabledBy: { type: string, format: uuid }
        disabledAt: { type: string, format: date-time }
        disabledReason: { type: string }
        createdAt: { type: string, format: date-time }
        expiringSoon: { type: boolean }
    ManageRedemptionOperation:
      type: object
      additionalProperties: false
      required: [action, operator, createdAt]
      properties:
        action: { type: string, enum: [generated, redeemed, disabled] }
        operator: { type: string }
        reason: { type: string }
        createdAt: { type: string, format: date-time }
    ManageRedemptionCodeDetail:
      allOf:
        - $ref: "#/components/schemas/ManageRedemptionCode"
        - type: object
          required: [operationHistory]
          properties:
            operationHistory:
              type: array
              items:
                $ref: "#/components/schemas/ManageRedemptionOperation"
    ManageGeneratedCode:
      type: object
      additionalProperties: false
      required: [id, fullCode, maskedCode]
      properties:
        id: { type: string, format: uuid }
        fullCode:
          type: string
          description: 仅平台管理员可见的完整兑换码
        maskedCode: { type: string }
    ManageRedemptionBatchCreatedData:
      type: object
      additionalProperties: false
      required: [batch, codes]
      properties:
        batch:
          $ref: "#/components/schemas/ManageRedemptionBatch"
        codes:
          type: array
          items:
            $ref: "#/components/schemas/ManageGeneratedCode"
    ManageRedemptionBatchExportData:
      type: object
      additionalProperties: false
      required: [filename, content, mediaType, format, count]
      properties:
        filename: { type: string }
        content:
          type: string
          description: >-
            下载正文。闲鱼格式中一行对应一个兑换码，固定为“浏览器打开领取并开始创作吧～：领取链接 ｜ 兑换码：完整兑换码”。
        csv:
          type: string
          description: 兼容旧管理端的 CSV 内容，仅 format=csv 时返回。
        mediaType: { type: string }
        format: { type: string, enum: [csv, xianyu] }
        count: { type: integer, minimum: 0 }
    ManageRedemptionCodeRevealData:
      type: object
      additionalProperties: false
      required: [id, fullCode]
      properties:
        id: { type: string, format: uuid }
        fullCode:
          type: string
          description: 仅平台管理员可见的完整兑换码
    ManageBulkMutationData:
      type: object
      additionalProperties: false
      required: [affected, skipped, failed]
      properties:
        affected: { type: integer, minimum: 0 }
        skipped: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
...
    ManageRedemptionBatchPage:
      allOf:
        - $ref: "#/components/schemas/ManagePageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ManageRedemptionBatch"
    ManageRedemptionCodePage:
      allOf:
        - $ref: "#/components/schemas/ManagePageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ManageRedemptionCode"
...
    ManageRedemptionBatchPageSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManageRedemptionBatchPage"
    ManageRedemptionBatchSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManageRedemptionBatch"
    ManageRedemptionBatchCreatedSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManageRedemptionBatchCreatedData"
    ManageGeneratedCodeListSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ManageGeneratedCode"
    ManageRedemptionBatchExportSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManageRedemptionBatchExportData"
    ManageRedemptionCodePageSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManageRedemptionCodePage"
    ManageRedemptionCodeDetailSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManageRedemptionCodeDetail"
    ManageRedemptionCodeRevealSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManageRedemptionCodeRevealData"
    ManageBulkMutationSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManageBulkMutationData"
...
    RedemptionCodeStatus:
      type: string
      enum: [unused, redeemed, expired, disabled]
    RedemptionBatchCreateInput:
      type: object
      additionalProperties: false
      required: [name, quantity, creditsPerCode, productCode]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        quantity: { type: integer, minimum: 1, maximum: 500 }
        creditsPerCode: { type: integer, minimum: 1, maximum: 100000 }
        productCode: { type: string, minLength: 1, maxLength: 100 }
        expiresAt: { type: [string, "null"], format: date-time }
        note: { type: string, maxLength: 1000 }
    RedemptionBatchUpdateInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
...